move setup.py parameters to setup.cfg
