+++ MANIFEST.in
@@ -3,15 +3,14 @@ include README.md
 include LICENSE.txt
 include MANIFEST.in
 
-
-
-
 include pyproject.toml
 include package.json
 include install.json
 include ts*.json
+include jupyter-config/ibis_vega_transform.json
 
 graft ibis_vega_transform
+graft ibis_vega_transform/labextension
 
 # Javascript files
 graft src
