+++ jupyter-config/jupyter_notebook_config.d/ibis_vega_transform.json
@@ -1,0 +1,7 @@
+{
+  "NotebookApp": {
+    "nbserver_extensions": {
+      "ibis_vega_transform": true
+    }
+  }
+}
