+++ jupyter-config/jupyter_server_config.d/ibis_vega_transform.json
@@ -1,0 +1,7 @@
+{
+  "ServerApp": {
+    "jpserver_extensions": {
+      "ibis_vega_transform": true
+    }
+  }
+}
