+++ pyproject.toml
@@ -1,3 +1,9 @@
 [build-system]
-requires = ["jupyter_packaging~=0.7.12", "jupyterlab~=2.0", "setuptools>=40.8.0", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["jupyter_packaging~=0.8.1"]
+build-backend = "jupyter_packaging.build_api"
+
+[tool.jupyter-packaging.builder]
+func = "jupyter_packaging.npm_builder"
+
+[tool.jupyter-packaging.build-args]
+build_cmd = "build:prod"
