+++ setup.cfg
@@ -1,0 +1,53 @@
+[metadata]
+name = ibis_vega_transform
+version = attr: ibis_vega_transform._version.__version__
+description = Evaluate Vega transforms using Ibis expressions.
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = Apache-2.0 License
+author = Ian Rose and Saul Shanabrook
+author_email = ian.rose@quansight.com
+url = https://github.com/Quansight/ibis-vega-transform
+platforms = Linux, Mac OS X, Windows
+keywords = jupyterlab-extension, ibis, vega
+classifiers =
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find:
+python_requires = >=3.7
+install_requires =
+    setuptools>=46.4.0
+    altair>=4.0.0
+    altair-transform
+    ibis-framework>=1.3.0
+    jaeger-client
+    jupyter_jaeger>=1.0.3
+    jupyterlab==3.*
+    jupyterlab_server
+    mypy_extensions
+    nbclassic>=0.2
+    opentracing
+    # NOTE: remove pymapd pinning when ibis v2 is ready
+    pymapd==0.24.*
+    # NOTE: current ibis version doesn't work with sqlalchemy 1.4
+    sqlalchemy<1.4
+    tornado
+    typing_extensions
+    vega_datasets
+
+[bdist_wheel]
+universal=1
+
+[manifix]
+known_excludes =
+    .git*
+    *.pyc
+    build
+    dist
+    .cache
+    .pytest_cache
